+++ .gitignore
@@ -5,6 +5,7 @@
 __pycache__/
 *.py[cod]
 *$py.class
+retnet301.pt
 
 # Environments
 .env
