+++ test.py
@@ -1,8 +1,11 @@
-from torchscale.architecture.config import RetNetConfig
-from hugging_faceify import hugging_faceify
+import torch
+from hugging_face_model import RetNetModel, TransformerModel
 
-config = RetNetConfig()
-config.save_pretrained('ret_config')
-ret = hugging_faceify('scripts/retnet301.pt', 'ret_config/config.json')
+retnet = RetNetModel(config='retnet_config/config.json')
+retnet.load_state_dict(torch.load('retnet_config/retnet301.pt'))
+print("RetNet worked!")
+transformer = TransformerModel(config='transformer_config/config.json')
+transformer.load_state_dict(torch.load('transformer_config/transformer.pt'))
+print("Transformer worked!")
