+++ transformer_config/config.json
@@ -1,0 +1,39 @@
+{
+  "activation_dropout": 0.1,
+  "activation_fn": "gelu",
+  "attention_dropout": 0.0,
+  "bert_init": false,
+  "checkpoint_activations": false,
+  "ddp_rank": 0,
+  "decoder_attention_heads": 8,
+  "decoder_embed_dim": 128,
+  "decoder_ffn_embed_dim": 1024,
+  "decoder_layers": 6,
+  "decoder_normalize_before": true,
+  "deepnorm": false,
+  "drop_path_rate": 0.0,
+  "dropout": 0.1,
+  "fsdp": true,
+  "layernorm_embedding": false,
+  "layernorm_eps": 1e-05,
+  "max_rel_pos": 0,
+  "max_target_positions": 1024,
+  "moe_eval_capacity_token_fraction": 0.25,
+  "moe_expert_count": 0,
+  "moe_freq": 0,
+  "moe_gating_use_fp32": true,
+  "moe_normalize_gate_prob_before_dropping": false,
+  "moe_second_expert_policy": "random",
+  "moe_top1_expert": false,
+  "multiway": false,
+  "no_output_layer": false,
+  "no_scale_embedding": true,
+  "rel_pos_buckets": 0,
+  "share_decoder_input_output_embed": false,
+  "subln": true,
+  "transformers_version": "4.36.2",
+  "use_xmoe": false,
+  "vocab_size": 100000,
+  "xpos_rel_pos": false,
+  "xpos_scale_base": 512
+}
